requirements.txt
fastapi
uvicorn
pydantic>=2.9.0
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-pinecone
pinecone-client
PyGithub